+++ Yank/experiment.py
@@ -433,7 +433,7 @@ class Experiment(object):
 
     def run(self, n_iterations=None):
         """
-        Run the experiment
+        Run the experiment.
 
         Runs until either the maximum number of iterations have been reached or the sampler
         for that phase reports its own completion (e.g. online analysis)
@@ -610,6 +610,7 @@ class ExperimentBuilder(object):
         'experiments_dir': 'experiments',
         'platform': 'fastest',
         'precision': 'auto',
+        'max_n_contexts': 3,
         'switch_experiment_interval': 0,
         'processes_per_experiment': None
     }
@@ -763,7 +764,7 @@ class ExperimentBuilder(object):
             # The cache has been already used. Empty it before switching platform.
             mmtools.cache.global_context_cache.empty()
             mmtools.cache.global_context_cache.platform = platform
-        mmtools.cache.global_context_cache.capacity = 3
+        mmtools.cache.global_context_cache.capacity = self._options['max_n_contexts']
 
         # Initialize and configure database with molecules, solvents and systems
         setup_dir = os.path.join(self._options['output_dir'], self._options['setup_dir'])
@@ -1234,13 +1235,24 @@ class ExperimentBuilder(object):
                                'number_of_iterations': integer_or_infinity,
                                'anisotropic_dispersion_cutoff': check_anisotropic_cutoff}
 
-        # Validate parameters
+        # Validate parameters.
         try:
             validated_options = utils.validate_parameters(options, template_options, check_unknown=True,
                                                           process_units_str=True, float_to_int=True,
                                                           special_conversions=special_conversions)
         except (TypeError, ValueError) as e:
             raise YamlParseError(str(e))
+
+        # Overwrite defaults.
+        defaults_to_overwrite = {
+            # With the analytical dispersion correction for alchemical atoms,
+            # the computation of the energy matrix becomes super slow.
+            'disable_alchemical_dispersion_correction': True,
+        }
+        for option, default_value in defaults_to_overwrite.items():
+            if option not in validated_options:
+                validated_options[option] = default_value
+
         return validated_options
 
     @classmethod